BJ-988: Refactor in favor of SRP
