+++ src/main/java/com/rbkmoney/threeds/server/config/rbkmoneyplatform/PreparationFlowScheduleConfig.java
@@ -5,7 +5,8 @@ import com.rbkmoney.damsel.three_ds_server_storage.InitRBKMoneyPreparationFlowRe
 import com.rbkmoney.threeds.server.config.properties.PreparationFlowDsProviderProperties;
 import com.rbkmoney.threeds.server.config.properties.PreparationFlowScheduleProperties;
 import com.rbkmoney.threeds.server.serializer.ThriftSerializer;
-import com.rbkmoney.threeds.server.service.RBKMoneyPreparationFlowScheduler;
+import com.rbkmoney.threeds.server.service.schedule.RBKMoneyPreparationFlowScheduler;
+import com.rbkmoney.threeds.server.service.schedule.SchedulatorService;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.ConfigurationProperties;
@@ -14,25 +15,33 @@ import org.springframework.context.annotation.Configuration;
 
 @Configuration
 @ConditionalOnProperty(name = "platform.mode", havingValue = "RBK_MONEY_PLATFORM")
-public class PreparationFlowSchedulerConfig {
+public class PreparationFlowScheduleConfig {
 
     @Bean
     public RBKMoneyPreparationFlowScheduler rbkMoneyPreparationFlowScheduler(
-            @Value("${client.three-ds-server-storage.card-ranges.url}") String threeDsServerStorageUrl,
-            PreparationFlowScheduleProperties preparationFlowScheduleProperties,
             PreparationFlowDsProviderProperties visaPreparationFlowProperties,
             PreparationFlowDsProviderProperties mastercardPreparationFlowProperties,
             PreparationFlowDsProviderProperties mirPreparationFlowProperties,
-            ThriftSerializer<InitRBKMoneyPreparationFlowRequest> preparationFlowRequestThriftSerializer,
-            SchedulatorSrv.Iface schedulatorClient) {
+            SchedulatorService schedulatorService) {
         return new RBKMoneyPreparationFlowScheduler(
-                threeDsServerStorageUrl,
-                preparationFlowScheduleProperties,
                 visaPreparationFlowProperties,
                 mastercardPreparationFlowProperties,
                 mirPreparationFlowProperties,
+                schedulatorService);
+    }
+
+    @Bean
+    public SchedulatorService schedulatorService(
+            @Value("${client.three-ds-server-storage.card-ranges.url}") String threeDsServerStorageUrl,
+            PreparationFlowScheduleProperties preparationFlowScheduleProperties,
+            ThriftSerializer<InitRBKMoneyPreparationFlowRequest> preparationFlowRequestThriftSerializer,
+            SchedulatorSrv.Iface schedulatorClient) {
+        return new SchedulatorService(
+                threeDsServerStorageUrl,
+                preparationFlowScheduleProperties,
                 preparationFlowRequestThriftSerializer,
                 schedulatorClient);
+
     }
 
     @Bean
